refactored python files for correct assignments of functions
