+++ requirements.txt
@@ -3,3 +3,4 @@ psycopg2-binary==2.9.9
 zeep==4.1.0
 mozilla_django_oidc==4.0.0
 PyJWT==2.8.0
+pytest
